INC-192 refactor: rename usersService to userService, move get posts query to userService
