+++ src/app/[locale]/(dashboard)/usersList/[id]/page.tsx
@@ -1,6 +1,3 @@
-import { BackToUsersListLink } from './_components'
-import { UserUploadedPhotos } from './_components/userUploadedPhotos'
-
-export default function UsersListPage() {
-  return <UserUploadedPhotos />
+export default function UserPage() {
+  return <>User page</>
 }
