+++ src/services/userService.ts
@@ -17,3 +17,16 @@ export const GET_USER = gql(`
     } 
   }
 `)
+
+export const GET_POSTS_BY_USER = gql(`
+  query GetPostsByUser($userId: Int!, $endCursorId: Int) {
+    getPostsByUser(userId: $userId, endCursorId: $endCursorId) {
+      pagesCount
+      items {
+        id
+        createdAt
+        url
+      }
+    } 
+  }
+`)
